style:current-limiting structs extract reuse
