+++ pkg/filter/sentinel/ratelimit/matcher_test.go
@@ -27,10 +27,14 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+import (
+	"github.com/apache/dubbo-go-pixiu/pkg/filter/sentinel"
+)
+
 func TestMatch(t *testing.T) {
 	config := mockConfig()
-	m := newMatcher()
-	m.load(config.Resources)
+	m := sentinel.NewMatcher()
+	m.Load(config.Resources)
 
 	tests := []struct {
 		give    string
@@ -70,7 +74,7 @@ func TestMatch(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.give, func(t *testing.T) {
-			res, ok := m.match(tt.give)
+			res, ok := m.Match(tt.give)
 			assert.Equal(t, res, tt.res)
 			assert.Equal(t, ok, tt.matched)
 		})
@@ -79,22 +83,22 @@ func TestMatch(t *testing.T) {
 
 func mockConfig() *Config {
 	c := Config{
-		Resources: []*Resource{
+		Resources: []*sentinel.Resource{
 			{
 				Name: "test-dubbo",
-				Items: []*Item{
-					{MatchStrategy: EXACT, Pattern: "/api/v1/test-dubbo/user"},
-					{MatchStrategy: REGEX, Pattern: "/api/v1/test-dubbo/user/*"},
+				Items: []*sentinel.Item{
+					{MatchStrategy: sentinel.EXACT, Pattern: "/api/v1/test-dubbo/user"},
+					{MatchStrategy: sentinel.REGEX, Pattern: "/api/v1/test-dubbo/user/*"},
 				},
 			},
 			{
 				Name: "test-http",
-				Items: []*Item{
-					{MatchStrategy: EXACT, Pattern: "/api/v1/http/foo"},
-					{MatchStrategy: EXACT, Pattern: "/api/v1/http/bar"},
+				Items: []*sentinel.Item{
+					{MatchStrategy: sentinel.EXACT, Pattern: "/api/v1/http/foo"},
+					{MatchStrategy: sentinel.EXACT, Pattern: "/api/v1/http/bar"},
 
-					{MatchStrategy: REGEX, Pattern: "/api/v1/http/foo/*"},
-					{MatchStrategy: REGEX, Pattern: "/api/v1/http/bar/*"},
+					{MatchStrategy: sentinel.REGEX, Pattern: "/api/v1/http/foo/*"},
+					{MatchStrategy: sentinel.REGEX, Pattern: "/api/v1/http/bar/*"},
 				},
 			},
 		},
